// oroasis/src/shared/enums/tool.ts
export enum EnumListTool {
    execute_command = "execute_command",
    read_file = "read_file",
    list_files = "list_files",
    write_file = "write_file",
    selection_in_file = "selection_in_file",
    replace_in_file = "replace_in_file",
    search_files = "search_files",
    use_mcp_tool = "use_mcp_tool"
}